modify schedule-lifetime for use in a list
Add some flexibility to allow building a list of schedule lifetimes.
This will accommodate changing the periodic behavior of a value or set
of values.

To simplify the schedule grouping, remove the choice between "always"
and "start-date-time".  This is unnecessary since the same goal can
be achieved with a start time in the past and an infinite duration.
Now the start-date-time can be used as a list key.

Through the use of an augment statement in a "uses" clause, this can
also allow independent variation of multiple attributes/values during
a single schedule lifetime.